Protect against literal interpretation (#137)
* Protect against literal interpretation

On msvc 2017 this seems to have been checking for existence of a flag called `_warning_flag`

* Stop checking for flag called "IN"

* Ensure compiler flag checks are not cached

Per the docs of CMake the call to CHECK_CXX_COMPILER_FLAG 
stores the result in an internal cache entry. In this loop we were 
basically skipping the "supported" check for all flags after the 
first one. Forcibly unsetting the variable before each check 
ensures that we actually check if each specific flag is supported